handle token expiration
error is inspected for token expiration warning, if present then user is
warned that token is expired and to clock-in again as if it were the
first time they were using the program. The token, last event time, and
last event summary files are deleted to reset program and auth token
flow.